Refactor into JaialogStore class, and all methods now take log file stems, rather than full paths
