+++ src/web/jdv/server/jaiabot_data_vision.py
@@ -7,10 +7,9 @@ import logging
 import os
 import math
 
-import jaialogs
+import jaialog_store
+import moos_messages
 import pyjaia.contours
-
-from jaia_h5 import JaiaH5FileSet
 
 # Arguments
 parser = argparse.ArgumentParser()
@@ -25,7 +24,7 @@ logging.basicConfig(level=logLevel)
 logging.getLogger('werkzeug').setLevel('WARN')
 
 # Setup the directory
-jaialogs.set_directory(os.path.expanduser(args.directory))
+jaialogStore = jaialog_store.JaialogStore(args.directory)
 
 app = Flask(__name__)
 
@@ -61,18 +60,18 @@ def getRoot():
 
 @app.route('/logs', methods=['GET'])
 def getLogs():
-    return JSONResponse(jaialogs.get_logs())
+    return JSONResponse(jaialogStore.getLogs())
 
 @app.route('/convert-if-needed', methods=['POST'])
 def convertLogs():
     log_names = request.json
-    return JSONResponse(jaialogs.convert_if_needed(log_names))
+    return JSONResponse(jaialogStore.convertIfNeeded(log_names))
 
 @app.route('/paths', methods=['GET'])
 def getFields():
     log_names = parse_log_filenames(request.args.get('log'))
     root_path = request.args.get('root_path')
-    return JSONResponse(jaialogs.get_fields(log_names, root_path))
+    return JSONResponse(jaialogStore.getFields(log_names, root_path))
 
 @app.route('/series', methods=['GET'])
 def getSeries():
@@ -80,7 +79,7 @@ def getSeries():
     series_names = request.args.get('path')
 
     try:
-        series = jaialogs.get_series(log_names, series_names)
+        series = jaialogStore.getSeries(log_names, series_names)
         return JSONResponse(series)
     except Exception as e:
         return JSONErrorResponse(str(e))
@@ -88,7 +87,7 @@ def getSeries():
 @app.route('/map', methods=['GET'])
 def getMap():
     log_names = parse_log_filenames(request.args.get('log'))
-    return JSONResponse(jaialogs.get_map(log_names))
+    return JSONResponse(jaialogStore.getMap(log_names))
 
 
 @app.route('/commands', methods=['GET'])
@@ -98,7 +97,7 @@ def getCommands():
     if log_names is None:
         return JSONErrorResponse("Missing log filename")
 
-    return JSONResponse(jaialogs.get_commands(log_names))
+    return JSONResponse(jaialogStore.getCommands(log_names))
 
 
 @app.route('/active-goal', methods=['GET'])
@@ -108,7 +107,7 @@ def getActiveGoals():
     if log_names is None:
         return JSONErrorResponse("Missing log filename")
 
-    return JSONResponse(jaialogs.get_active_goals(log_names))
+    return JSONResponse(jaialogStore.getActiveGoals(log_names))
 
 
 @app.route('/task-packet', methods=['GET'])
@@ -118,7 +117,7 @@ def getTaskPackets():
     if log_names is None:
         return JSONErrorResponse("Missing log filename")
 
-    return JSONResponse(jaialogs.get_task_packet_dicts(log_names))
+    return JSONResponse(jaialogStore.getTaskPacketDicts(log_names))
 
 
 @app.route('/moos', methods=['GET'])
@@ -131,7 +130,7 @@ def getMOOSMessages():
     if log_names is None:
         return JSONErrorResponse("Missing log filename")
 
-    return Response(jaialogs.get_moos_messages(log_names, t_start, t_end), mimetype='text/csv')
+    return Response(moos_messages.get_moos_messages(log_names, t_start, t_end), mimetype='text/csv')
 
 
 @app.route('/depth-contours', methods=['GET'])
@@ -142,7 +141,7 @@ def getDepthContours():
     if log_names is None:
         return JSONErrorResponse("Missing log filename")
 
-    taskPackets = jaialogs.get_task_packet_dicts(log_names)
+    taskPackets = jaialogStore.getTaskPacketDicts(log_names)
     return JSONResponse(pyjaia.contours.taskPacketsToContours(taskPackets))
